Update openwrt.yaml
Changed this name from "OpenWRT" to "openwrt-15.05-x86-kvm_guest-combined-ext4" because of the image included in the sample local.conf file will be uploaded to Glance with the following name. As proposed in the following fix:  https://github.com/openstack/tacker/compare/master...Aqsamm:patch-1
The image will then be pre-installed in the devstack installation.